requirements.txt
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchmetrics==1.2.0
segmentation_models_pytorch==0.5.0
tqdm==4.67.1
opencv-python==4.11.0.86
pillow==11.0.0
numpy==2.1.2
pandas==2.2.3
imageio==2.37.0